+++ apps/vmq_server/src/vmq_schema.erl
@@ -1,3 +1,16 @@
+%% Copyright 2018 Erlio GmbH Basel Switzerland (http://erl.io)
+%%
+%% Licensed under the Apache License, Version 2.0 (the "License");
+%% you may not use this file except in compliance with the License.
+%% You may obtain a copy of the License at
+%%
+%%     http://www.apache.org/licenses/LICENSE-2.0
+%%
+%% Unless required by applicable law or agreed to in writing, software
+%% distributed under the License is distributed on an "AS IS" BASIS,
+%% WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+%% See the License for the specific language governing permissions and
+%% limitations under the License.
 -module(vmq_schema).
 
 -export([translate_listeners/1]).
@@ -205,7 +218,8 @@ translate_listeners(Conf) ->
      {vmq, DropUndef(VMQ)},
      {vmqs, DropUndef(VMQS)},
      {http, DropUndef(HTTP)},
-     {https, DropUndef(HTTPS)}].
+     {https, DropUndef(HTTPS)}
+    ].
 
 extract(Prefix, Suffix, Val, Conf) ->
     Mappings = ["max_connections", "nr_of_acceptors", "mountpoint"],
@@ -234,11 +248,12 @@ extract(Prefix, Suffix, Val, Conf) ->
     [begin
          {ok, Addr} = inet:parse_address(StrAddr),
          Prefix4 = lists:flatten([Prefix, ".", Name, ".", Suffix]),
-         V = Val(Name, cuttlefish:conf_get(Prefix4, Conf, Default),
-                 Val(Name, RootDefault,
-                     Val(Name, RootDefault, undefined))),
+         V1 = Val(Name, RootDefault, undefined),
+         V2 = Val(Name, RootDefault,V1),
+         V3 = Val(Name, cuttlefish:conf_get(Prefix4, Conf, Default),V2),
+
          AddrPort = {Addr, Port},
-         {AddrPort, {list_to_atom(Suffix), V}}
+         {AddrPort, {list_to_atom(Suffix), V3}}
      end
      || {[_, _, Name], {StrAddr, Port}} <- lists:filter(
                                              fun({K, _V}) ->